Attempt to organise the sandbox module
