+++ sandbox/README.md
@@ -1,10 +1,3 @@
-# SANBOX PROJECT
+# SANDBOX PROJECT
+
 This project is used to test the library.
-
-## Using the sandbox project.
-First go to the root of the project and run:
-`scala-cli --power publish local .`
-
-Then go to the sandbox directory and run:
-
-`bash run.sh`
